Redirect to correct action creators documentation
I found this resource and excellent intro to Gatsby, thanks! I wanted to give back by fixing this small bug I found whilst trying to access more of the action creator documentation, it just seems you need the domain of Gatsby or else it was going to that page on your site.